requirements.txt
playwright
psycopg2-binary
requests
python-dotenv